I added the media queries tp the logo and the table
Can you check if it works on your end. Because when I put the change to the table in the codeine it works but not in the browser and logo width and height numbers are not definite yet. I see the changes on the index page for logo but not on the rest of the pages when change the screen sizes.